refactor(asset): moving split implementation to schema
