+++ modules/splits/internal/mappable/mappable.go
@@ -1,0 +1,35 @@
+// Copyright [2021] - [2022], AssetMantle Pte. Ltd. and the code contributors
+// SPDX-License-Identifier: Apache-2.0
+
+package mappable
+
+import (
+	"github.com/cosmos/cosmos-sdk/codec"
+
+	"github.com/AssetMantle/modules/modules/splits/internal/key"
+	"github.com/AssetMantle/modules/schema/helpers"
+	"github.com/AssetMantle/modules/schema/ids/base"
+	"github.com/AssetMantle/modules/schema/mappables"
+	codecUtilities "github.com/AssetMantle/modules/utilities/codec"
+)
+
+type mappable struct {
+	mappables.Split
+}
+
+var _ helpers.Mappable = (*mappable)(nil)
+
+func (mappable mappable) GetKey() helpers.Key {
+	return key.NewKey(base.NewSplitID(mappable.Split.GetOwnerID(), mappable.Split.GetOwnableID()))
+}
+func (mappable) RegisterCodec(codec *codec.Codec) {
+	codecUtilities.RegisterModuleConcrete(codec, mappable{})
+}
+
+func NewMappable(split mappables.Split) helpers.Mappable {
+	return mappable{Split: split}
+}
+
+func Prototype() helpers.Mappable {
+	return mappable{}
+}
